Do not create duplicate clusters
The old Carina API would "ignore" requests to create a cluster if
there was already a cluster with the same name (because the primary key
was the cluster name). Now that clusters can exist with the same name,
we need to check first if it exists, and skip the call to create.